% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_POSIXct.R
\name{from_POSIXct}
\alias{from_POSIXct}
\title{Convert POSIXct Columns to ISO-8601 Character}
\usage{
from_POSIXct(data)
}
\arguments{
\item{data}{a data.frame (expected to contain POSIXct columns)}
}
\value{
a data.frame, whose POSIXct columns have been converted to ISO-8601 character columns
}
\description{
Convert POSIXct Columns to ISO-8601 Character
}
\examples{
from_POSIXct(data.frame(a = Sys.time()))
}
