% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read Data}
\usage{
read_data(
  path = Sys.getenv("DATA_HOME"),
  resource = NULL,
  file,
  delim = ",",
  col_types = NULL,
  verbose = FALSE
)
}
\arguments{
\item{path}{A path where to read the data (see details).}

\item{resource}{An optional resource name (used as sub folder inside path).}

\item{file}{The name of the file to read (with extension).}

\item{delim}{Single character used to separate fields within a record.}

\item{col_types}{See read_delim function from readr package.}

\item{verbose}{A logical whether additional trace should be sent to the console.}
}
\value{
A tibble, output value of the read_delim() call or NULL if data can't be read.
}
\description{
Read Data
}
\details{
By default, path takes its value from the DATA_HOME environment variable.
If it's not set in the environment and no value has been provided, function will be stopped (and return NULL).
When NULL, it will be ignored (assuming file contains the full path to the file)
}
\examples{
\dontrun{
foo <- read_data(file = "my_data.csv")
}
}
