% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write Data}
\usage{
write_data(
  x,
  path = Sys.getenv("DATA_HOME"),
  resource = NULL,
  file,
  delim = ",",
  verbose = FALSE
)
}
\arguments{
\item{x}{A data frame or tibble to write to disk.}

\item{path}{A path where to write the data (see details).}

\item{resource}{An optional resource name (used as sub folder inside path).}

\item{file}{The name of the file to write (with extension)}

\item{delim}{Single character used to separate fields within a record.}

\item{verbose}{If TRUE, additional messages will be sent to the console}
}
\value{
returns the input x invisibly, or FALSE when it fails
}
\description{
Write Data
}
\details{
By default, path takes its value from the DATA_HOME environment variable.
If it's not set in the environment and no value has been provided, function will be stopped (and return FALSE).
If the corresponding path does not exist, it won't be created - to avoid creating data inside a container if the mount has not been done for example.
When NULL, no check will be done.

Resource is an optional parameter to allow the use of sub folder inside path.
If not NULL, it will be concatenated to path (and created inside path if it does not exist)
When path is NULL, it will be ignored.
}
\examples{
\dontrun{
write_data(foo, file = "my_data.csv")
}
}
