% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load Data From Various Connectors}
\usage{
load_data(..., verbose = FALSE, data.frame = FALSE)
}
\arguments{
\item{...}{arguments to pass to connector (see details)}

\item{verbose}{a logical (default = FALSE) to indicate if additional traces should be sent to the console}

\item{data.frame}{a logical (default = FALSE) if a data.frame should be returned (instead of a tibble)}
}
\value{
the output of the connector function
}
\description{
Load Data From Various Connectors
}
\details{
The load_data function is a wrapper around connector functions.
It offers a single point of access to the package data persistence features.

Arguments passed to ... depend on the desired connector:
\itemize{
\item file: \link[iker]{read_data}
}
}
\examples{
\dontrun{
# -- file connector
load_data(file = "my_data.csv", path = getwd())
}
}
